define first level quality
